Linking Sectors
A lot needs to be done to link sectors. Jumpgates are hard-coded to lead
to certain other jumpgates, and so each gate should be self-contained to
determine where it leads, and to transfer the user to the proper
destination.

Also, laying the foundation for loading data and placing the player
properly in their last location, and saving the current game to the
current save ID (saving multiple to new files will come later)